# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseLoginView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoginView"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseObject": {
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponsePageResultUserView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResultUserView"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseUserView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserView"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "captchaToken": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 72,
            "minLength": 0,
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "captchaToken",
          "password",
          "username"
        ],
        "type": "object"
      },
      "LoginView": {
        "properties": {
          "token": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
...
      "PageResultUserView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserView"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "password": {
            "maxLength": 64,
            "minLength": 10,
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
...
      "TargetVersion": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version"
        ],
        "type": "object"
      },
...
      "UserRequest": {
        "properties": {
          "departmentId": {
            "format": "int64",
            "type": "integer"
          },
          "email": {
            "format": "email",
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "nickname": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "phone": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "postIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array",
            "uniqueItems": true
          },
          "roleIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array",
            "uniqueItems": true
          },
          "username": {
            "minLength": 1,
            "pattern": "[a-zA-Z0-9_]{3,32}",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "nickname",
          "roleIds",
          "username"
        ],
        "type": "object"
      },
      "UserStatusRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TargetVersion"
            },
            "maxItems": 100,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "type": "object"
      },
      "UserView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "departmentId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "departmentName": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "postIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array",
            "uniqueItems": true
          },
          "roleIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array",
            "uniqueItems": true
          },
          "roleNames": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "username": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "createdAt",
          "departmentId",
          "departmentName",
          "email",
          "enabled",
          "id",
          "nickname",
          "phone",
          "postIds",
          "roleIds",
          "roleNames",
          "username",
          "version"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "通过滑块验证与登录接口取得的一次性会话令牌",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "统一响应、动作权限、数据范围与乐观锁均由服务端执行。关闭模块不开放接口。",
    "title": "Mayday API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLoginView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [],
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/system/user-statistics": {
      "get": {
        "operationId": "UserStatisticsController_statistics",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-statistics-controller"
        ]
      }
    },
    "/api/system/users": {
      "get": {
        "operationId": "UserController_list",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultUserView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:view')"
      },
      "post": {
        "operationId": "UserController_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:create')"
      }
    },
    "/api/system/users/export": {
      "get": {
        "operationId": "UserController_export",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "departmentId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultUserView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:export') and @access.has('users:view')"
      }
    },
    "/api/system/users/status": {
      "put": {
        "operationId": "UserController_status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:update')"
      }
    },
    "/api/system/users/{id}": {
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:delete')"
      },
      "put": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:update')"
      }
    },
    "/api/system/users/{id}/password": {
      "put": {
        "operationId": "UserController_reset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user-controller"
        ],
        "x-authorization": "@access.has('users:reset')"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}